Mini refactor
