+++ src/extension/index.ts
@@ -7,7 +7,7 @@ const ext = new RunmeExtension()
 export async function activate (context: ExtensionContext) {
   console.log('[Runme] Activating Extension')
   try {
-    await ext.initialise(context)
+    await ext.initialize(context)
     console.log('[Runme] Extension successfully activated')
   } catch (err: any) {
     console.log(`[Runme] Failed to initialise the extension ${err.message}`)
