+++ src/extension/notebook.ts
@@ -1,9 +1,6 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
 import {
   NotebookSerializer, ExtensionContext, Uri, workspace, NotebookData, NotebookCellData, NotebookCellKind,
-  window,
-  CancellationToken
-} from 'vscode'
+  window} from 'vscode'
 
 import type { WasmLib } from '../types'
 
@@ -15,7 +12,7 @@ import { normalizeLanguage, verifyCheckedInFile } from './utils'
 declare var globalThis: any
 
 // const CODE_REGEX = /```(\w+)?\n[^`]*```/g
-const DEFAULT_LANG_ID = 'text'
+export const DEFAULT_LANG_ID = 'text'
 
 export class Serializer implements NotebookSerializer {
   private readonly ready: Promise<Error | void>
@@ -190,119 +187,4 @@ export class Serializer implements NotebookSerializer {
   }
 }
 
-export class NewSerializer implements NotebookSerializer {
-  private readonly wasmReady: Promise<void>
-  private readonly languages: Languages
 
-  constructor(private context: ExtensionContext) {
-    this.languages = Languages.fromContext(this.context)
-    this.wasmReady = this.#initWasm()
-  }
-
-  async #initWasm() {
-    const go = new globalThis.Go()
-    const wasmUri = Uri.joinPath(this.context.extensionUri, 'wasm', 'runme.wasm')
-    const wasmFile = await workspace.fs.readFile(wasmUri)
-    return WebAssembly.instantiate(wasmFile, go.importObject).then(
-      (result) => { go.run(result.instance) },
-      (err: Error) => {
-        console.error(`[Runme] failed initialising WASM file: ${err.message}`)
-        throw err
-      }
-    )
-  }
-
-  public async serializeNotebook(data: NotebookData, token: CancellationToken): Promise<Uint8Array> {
-    try {
-      await this.wasmReady
-      const { Runme } = globalThis as WasmLib.New.Serializer
-
-      const notebook = JSON.stringify(data)
-      const markdown = await Runme.serialize(notebook)
-
-      const encoder = new TextEncoder()
-      const encoded = encoder.encode(markdown)
-
-      return encoded
-    } catch (err: any) {
-      console.error(err)
-      throw err
-    }
-  }
-
-  public async deserializeNotebook(content: Uint8Array, token: CancellationToken): Promise<NotebookData> {
-    let notebook: WasmLib.New.Notebook
-    try {
-      await this.wasmReady
-      const { Runme } = globalThis as WasmLib.New.Serializer
-
-      const markdown = content.toString()
-
-      notebook = await Runme.deserialize(markdown)
-
-      if (!notebook || (notebook.cells ?? []).length === 0) {
-        return this.#printCell('⚠️ __Error__: no cells found!')
-      }
-    } catch (err: any) {
-      return this.#printCell(
-        '⚠️ __Error__: document could not be loaded' +
-        (err ? `\n<small>${err.message}</small>` : '') +
-        '.<p>Please report bug at https://github.com/stateful/vscode-runme/issues' +
-        ' or let us know on Discord (https://discord.gg/BQm8zRCBUY)</p>'
-      )
-    }
-
-    try {
-      const cells = notebook.cells ?? []
-      notebook.cells = await Promise.all(cells.map(elem => {
-        if (elem.kind === NotebookCellKind.Code && elem.value && !elem.languageId) {
-          const norm = Serializer.normalize(elem.value)
-          return this.languages.guess(norm, PLATFORM_OS).then(guessed => {
-            elem.languageId = guessed
-            return elem
-          })
-        }
-        return Promise.resolve(elem)
-      }))
-    } catch (err: any) {
-      console.error(`Error guessing snippet languages: ${err}`)
-    }
-
-    const cells = NewSerializer.revive(notebook)
-    return new NotebookData(cells)
-  }
-
-  protected static revive(notebook: WasmLib.New.Notebook) {
-    return notebook.cells.reduce((accu, elem, id) => {
-      let cell: NotebookCellData
-      const isSupported = Object.keys(executor).includes(elem.languageId ?? '')
-
-      if (elem.kind === NotebookCellKind.Code && isSupported) {
-        cell = new NotebookCellData(
-          NotebookCellKind.Code,
-          elem.value,
-          elem.languageId || DEFAULT_LANG_ID
-        )
-      } else {
-        cell = new NotebookCellData(
-          NotebookCellKind.Markup,
-          elem.value,
-          'markdown'
-        )
-      }
-
-      // todo(sebastian): incomplete
-      cell.metadata = {...elem.metadata, id }
-      accu.push(cell)
-
-      return accu
-    }, <NotebookCellData[]>[])
-  }
-
-  #printCell(content: string, languageId = 'markdown') {
-    return new NotebookData([
-      new NotebookCellData(NotebookCellKind.Markup, content, languageId)
-    ])
-  }
-
-}
